updated server folder structure
