+++ server-api/src/lib/graphql/index.js
@@ -1,0 +1,7 @@
+import { GraphQLSchema } from 'graphql';
+import rootQuery from './query';
+
+module.exports = new GraphQLSchema({
+  query: rootQuery
+  // mutation: Mutations,
+});
