+++ server-api/src/server/graphqlConfig.js
@@ -1,13 +1,12 @@
 import graphqlHTTP from 'express-graphql';
 
-import schema from '../lib/graphql/schema';
+import schema from '../lib/graphql';
 
-export default (app) => {
+export default app =>
   app.use(
     '/api/graphql',
     graphqlHTTP({
       schema,
-      graphiql: true,
-    }),
+      graphiql: true
+    })
   );
-};
